refactor: 지불금액 Money객체로 포장
